pyproject.toml
# FX Data Platform Project Configuration
# This file defines project metadata, dependencies, and tool configurations

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fx-data-platform"
version = "0.1.0"
description = "Real-time FX transaction processing with ML-based anomaly detection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "FX Data Team", email = "data@example.com"},
]
keywords = [
    "data-engineering",
    "real-time-processing",
    "spark",
    "kafka",
    "ml-ops",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Core data processing
    "pyspark==3.5.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    
    # Streaming
    "confluent-kafka>=2.3.0",
    
    # Data validation and quality
    "great-expectations>=0.18.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.20.0",
    
    # ML and feature store
    "mlflow>=2.10.0",
    "feast>=0.38.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.1.4,<3.0.0",
    "pymc>=5.10.0",
    
    # Database and ORM
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    
    # Airflow
    "apache-airflow>=3.0.0",
    "apache-airflow-providers-amazon>=9.0.0",
    "apache-airflow-providers-postgres>=6.0.0",
    
    # Monitoring and logging
    "prometheus-client>=0.19.0",
    "python-json-logger>=2.0.0",
    "sentry-sdk>=1.39.0",
    
    # API and web
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    
    # Utilities
    "click>=8.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "structlog>=23.3.0",
    "boto3>=1.34.0",
    "moto>=4.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "testcontainers>=3.7.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "jupyterlab>=4.0.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "testcontainers>=3.7.0",
    "moto>=4.2.0",
    "fakeredis>=2.20.0",
]

[tool.setuptools]
packages = ["ingestion", "etl", "orchestration", "ml", "monitoring", "scripts"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | tests/.*?/setup\.py
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["ingestion", "etl", "orchestration", "ml", "monitoring", "scripts"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "A", "C4", "PIE", "T20", "PT"]
ignore = ["E501", "D100", "D104", "D105", "D107"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "confluent_kafka.*",
    "pyspark.*",
    "feast.*",
    "mlflow.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "slow: slow tests",
    "e2e: end-to-end tests",
]

[tool.coverage.run]
branch = true
source = ["ingestion", "etl", "orchestration", "ml", "monitoring", "scripts"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if 0:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
